api/requirements.txt
redis==2.10.6
Flask_RESTful==0.3.6
Flask==0.12.2
pyzmq==17.0.0b3
